Committing changes.

+++ mercado-additional/src/main/java/org/marceloleite/mercado/additional/filter/OldTradeTypeFilter.java
@@ -7,7 +7,7 @@ import org.marceloleite.mercado.commons.interfaces.Filter;
 import org.marceloleite.mercado.databasemodel.Trade;
 import org.marceloleite.mercado.databasemodel.TradeType;
 
-public class OldTradeTypeFilter implements Filter<Map<Integer, Trade>> {
+public class OldTradeTypeFilter implements Filter<Map<Long, Trade>> {
 
 	private TradeType type;
 
@@ -17,13 +17,10 @@ public class OldTradeTypeFilter implements Filter<Map<Integer, Trade>> {
 	}
 
 	@Override
-	public Map<Integer, Trade> filter(Map<Integer, Trade> trades) {
-		return trades.entrySet()
-			.stream()
-			.filter(entry -> type.equals(entry.getValue()
-				.getTradeType()))
-			.map(entry -> entry.getValue())
-			.collect(Collectors.toConcurrentMap(Trade::getId, trade -> trade, (oldTrade, newTrade) -> newTrade));
+	public Map<Long, Trade> filter(Map<Long, Trade> trades) {
+		return trades.entrySet().stream().filter(entry -> type.equals(entry.getValue().getTradeType()))
+				.map(entry -> entry.getValue())
+				.collect(Collectors.toConcurrentMap(Trade::getId, trade -> trade, (oldTrade, newTrade) -> newTrade));
 	}
 
 }
